refactor: rename test-utils.test.ts to test-utils.ts to remove false skips
