+++ src/scheduler/ReminderProcessor.channel.test.ts
@@ -12,7 +12,7 @@ import {
   APP_USER,
   ROOM,
   BASE_REMINDER,
-} from './ReminderProcessor.test-utils.test.ts';
+} from './ReminderProcessor.test-utils.ts';
 
 describe('ReminderProcessor – channel targets', () => {
   it('sends a message to a channel target', async () => {
